[ranking]: compute relevant_score
