+++ db/schema.rb
@@ -13,7 +13,7 @@
 
 ActiveRecord::Schema.define(version: 20150307103111) do
 
-  create_table "delayed_jobs", force: true do |t|
+  create_table "delayed_jobs", force: :cascade do |t|
     t.integer  "priority",   default: 0, null: false
     t.integer  "attempts",   default: 0, null: false
     t.text     "handler",                null: false
@@ -29,7 +29,7 @@ ActiveRecord::Schema.define(version: 20150307103111) do
 
   add_index "delayed_jobs", ["priority", "run_at"], name: "delayed_jobs_priority"
 
-  create_table "items", force: true do |t|
+  create_table "items", force: :cascade do |t|
     t.integer  "item_type_cd",                 null: false
     t.datetime "created_at",                   null: false
     t.boolean  "deleted",      default: false, null: false
@@ -54,7 +54,7 @@ ActiveRecord::Schema.define(version: 20150307103111) do
   add_index "items", ["story_id"], name: "index_items_on_story_id"
   add_index "items", ["updated_at"], name: "index_items_on_updated_at"
 
-  create_table "users", force: true do |t|
+  create_table "users", force: :cascade do |t|
     t.string   "username",               null: false
     t.text     "about"
     t.integer  "karma"
